Split session_manager.rs into per-decorator files
